# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: KIWAME API REST
  description: API for the management of KIWAME 
...
securityDefinitions:
  ApiKey:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
    description: "Security sheme migrated from security policy"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.


     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.


     The following syntax must be used in the 'Authorization' header :

       Bearer xxxxxx.yyyyyyy.zzzzzz


# during dev, should point to your local machine
#host: 80.58.61.69  #PRODUCCIÓN
#host: 138.4.47.33:8163  #UPM DEV
#host: localhost:8090  #LOCAL DEV
# basePath prefixes all resource paths
basePath: / #dds
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /messages/count:
    x-swagger-router-controller: message
    get:
      tags:
        - messages
      summary: Get number of messages  
      description: Get the logged user number of messages
      operationId: countMessages
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - patient
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: notifications retrieved
          schema:
            $ref: "#/definitions/MessagesCountResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"  
  /messages/{user}:
    x-swagger-router-controller: message
    post:
      tags:
        - messages
      summary: post message   
      description: post message to a conected user
      operationId: newMessage
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - patient
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user
          in: path
          description: The user nickname of the receiver
          type: string
          required: true
        - name: body
          in: body
          description: the actual message
          schema:
            required:
              - message
            properties:
              message:
                type: string
                example: message
          required: true
      responses:
        "201":
          description: Message sent
          schema:
            $ref: "#/definitions/ServerResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Receiver Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    get:
      tags:
        - messages
      summary: Retrieve messages  
      description: Retreive messages from connected users
      operationId: getMessages
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - patient
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user
          in: path
          description: The name of the user to get all messages
          type: string
          required: true
        - name: next
          in: query
          description: the id for retrieve the next page of results
          type: number
          required: false
      responses:
        "200":
          description: patients retrieved
          schema:
            $ref: "#/definitions/MessagesResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - messages
      summary: Mark as read 
      description: Mark message as read all messages from an user
      operationId: markMessageAsRead
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - patient
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: user
          in: path
          description: The name  of the user to mark as read his messages
          type: string
          required: true
      responses:
        "200":
          description: messages set as viewed
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /key/all:
    x-swagger-router-controller: key
    get:
      tags:
        - keys
      summary: get all keys
      description: Retrieve all keys from the user
      operationId: findAllKeys
      security:
        - Bearer: []
      x-security-scopes:
        - admin
        - user
      consumes:
        - application/json
      produces:
        - application/json

      responses:
        "200":
          description: Keys retrieved
          schema:
            $ref: "#/definitions/KeysResponse"
        "404":
          description: Keys Not found
          schema:
            $ref: "#/definitions/NotFoundResponse"

        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /key/{id}:
    x-swagger-router-controller: key
    delete:
      tags:
        - keys
      summary: Delete  key
      description: Create a new key in the DB
      operationId: deleteKey
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the key to delete
          type: string
          required: true
      responses:
        "204":
          description: deleted Success
        "404":
          description: Key Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - keys
      summary: Toggle enabled
      description: Update the key to enable or disable
      operationId: toggleEnabledKey
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the key to update
          type: string
          required: true

      responses:
        "200":
          description: Key Updated
          schema:
            # a pointer to a definition
            $ref: "#/definitions/KeyResponse"
        "404":
          description: Key Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /rule/{id}:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: findById
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule retrieves
          schema:
            $ref: "#/definitions/RuleResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: toogleEnabled
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule Enabled/Disabled
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    delete:
      tags:
        - rules
      summary: Delete rule
      description: Delete rule
      operationId: deleteRule
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to delete
          type: string
          required: true
      responses:
        "200":
          description: Rule removed
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
definitions:
...
  KeyObject:
    required:
      - name
    properties:
      _id:
        type: string
      name:
        type: string
        example: name
      enabled:
        type: boolean
        example: false
      key:
        type: string
        example: 54cd66-69d3e0fb-asa1b302e54e8
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
...
  MessageObject:
    required:
      - id
      - sender
      - message
      - viewed
      - created
    properties:
      id:
        type: string
        example: "the id of the message"
      sender:
        type: boolean
        example: true
      message:
        type: string
        example: "User has a new message"
      viewed:
        type: boolean
        example: false
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
...
  RuleRetrievedObject:
    required:
      - type
      - condition
      - value
      - notificationType
    properties:
      id:
        type: string
      patient:
        type: string
        example: nickname
      type:
        type: number
        example: 1
      measureType:
        type: number
        example: 2
      nutrientType:
        type: number
        example: 1
      condition:
        type: number
        example: 1
      value:
        type: number
        example: 10
      period:
        type: number
        minimum: 1
        maximum: 30
        example: 1
      notificationType:
        type: number
        example: 1
      enabled:
        type: boolean
        example: true
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
      updated:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
...
  NotFoundResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Resource Not found"
      statusCode:
        type: number
        example: 404
...
  KeysResponse:
    required:
      - message
      - statusCode
      - keys
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      keys:
        type: array
        items:
          $ref: "#/definitions/KeyObject"
  KeyResponse:
    required:
      - message
      - statusCode
      - key
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      key:
        $ref: "#/definitions/KeyObject"
...
  MessagesCountResponse:
    required:
      - message
      - statusCode
      - messages
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      messages:
        type: number
        example: 0
  MessagesResponse:
    required:
      - message
      - statusCode
      - messages
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      messages:
        type: array
        items:
          $ref: "#/definitions/MessageObject"
      next:
        type: string
        example: "/messages/profesional?next=1"
...
  RuleResponse:
    required:
      - message
      - statusCode
      - rule
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      rule:
          $ref: "#/definitions/RuleRetrievedObject"
  ServerResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Success"
      statusCode:
        type: number
        example: 200
...
  BadRequestResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Bad Request"
      statusCode:
        type: number
        example: 400
  AccessDeniedResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Access Denied"
      statusCode:
        type: number
        example: 403
...
  ServerErrorResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Internal Server Error"
      statusCode:
        type: number
        example: 500